Moved test file to the component folder.
